# swagger.json
{
    "swagger": "2.0",
    "info": {
        "descripton": "This is eKOSORA. The next URUBUTO but better",
        "version": "1.0.0",
        "title": "Academics management system",
        "contant": {
            "email": "ishimvainqueur@gmail.com"
        },
        "licence": {
            "name": "MIT License"
        }
    },
    "host": "localhost:80",
    "basePath": "/",
...
    "schemes": ["http"],
    "paths": {
...
        "student/register": {
            "post": {
                "tags": ["student"],
                "summary": "Add student to the database",
                "description": "This route can only be used by educators with the respective cookies to add new students to the database",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "New student info object",
                    "schema": {
                        "$ref": "#/definitions/newStudentBody"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successfully added a new student to the database",
                        "schema": {
                            "$ref": "#/definitions/newStudent200Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong in the server or the database",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/view": {
            "get": {
                "tags": ["student"],
                "summary": "Get all the students in the database",
                "description": "A full list of all students saved in the database and some info about them",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Successfully fetched all the students",
                        "schema": {
                            "$ref": "#/definitions/viewStudents200Response"
                        }
                    },
                    "500": {
                        "description": "In case an error occured while trying to fetch all the students",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "student/edit?id": {
            "post": {
                "tags": ["student"],
                "summary": "Edit a student's info",
                "description": "You can edit a student's personal info like names, class, or email",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "An object containing the new student info",
                    "schema": {
                        "$ref": "#/definitions/editStudentBody"
                    }
                    
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "required": true,
                        "description": "The _id of the student whose information needs to be changed"
                        
                    }],
                "responses": {
                    "200": {
                        "description": "Succesfully edited the student's info",
                        "schema": {
                            "$ref": "#/definitions/editStudent200Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request.\nPossible mistakes include: not including the student's id in the query parameters, passing an invalid id, ...",
                        "schema": {
                            "$ref": "#/definitions/editStudent400Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "student/addRecord": {
            "post": {
                "tags": ["student"],
                "summary": "Add record to all students",
                "description": "Add a record [Eg: Quiz, test, exam] to all the students of a specified class and specified subject",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "New record body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/addRecordBody"
                    }

                }],
                "responses": {
                    "200": {
                        "description": "Successfully added the new record to all students of the class",
                        "schema": {
                            "$ref": "#/definitions/addRecord200Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/updateMark":{
            "post": {
                "tags": ["student"],
                "summary": "Update one student's mark",
                "description": "You can change only one student's mark on a certain record",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "Updated mark object",
                    "schema":{
                        "$ref": "#/definitions/updateMarkBody"
                    }

                }],
                "responses": {
                    "200": {
                        "description": "Succesfully updated the student's mark",
                        "schema": {
                            "$ref": "#/definitions/updateMark200Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/updateForMany": {
            "post": {
                "tags": ["student"],
                "summary": "Update many student's marks at once",
                "description": "Add or remove a specified number of marks on a specified record for multiple students",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "The new marks object",
                    "schema": {
                        "$ref": "#/definitions/updateForManyBody"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successfully updated the students' marks",
                        "schema": {
                            "$ref": "#/definitions/updateForMany200Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/getRecords": {
            "get": {
                "tags": ["student"],
                "summary": "Get all records of a class",
                "description": "Returns all the records of a certain class in a certain subject",
                "produces": ["application/json"],
                "parameters": [{
                    "in": "query",
                    "name": "year",
                    "required": true,
                    "description": "The academic year of the class",
                    "schema": {
                        "$ref": "#/definitions/getRecordsYearQuery"
                    }
                    }, 
                    {
                        "in": "query",
                        "name": "class",
                        "required": true,
                        "description": "The class label",
                        "schema": {
                            "$ref": "#/definitions/getRecordsClassQuery"
                        }
                    },
                    {
                        "in": "query",
                        "name": "subject",
                        "required": true,
                        "description": "The class label",
                        "schema": {
                            "$ref": "#/definitions/getRecordsSubjectQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all the records in the specified class and subject",
                        "schema": {
                            "$ref": "#/definitions/getRecords200Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/deleteRecord": {
            "delete": {
                "tags": ["student"],
                "summary": "Delete a record",
                "description": "Delete a record from all students who have that particular records",
                "produces":[ "application/json" ],
                "parameters": [{
                    "in": "query",
                    "name": "_id",
                    "description": "The _id of the record in question",
                    "schema": {
                        "$ref": "#/definitions/deleteRecordIdQuery"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the record",
                        "schema": {
                            "$ref": "#/definitions/deleteRecord200Response"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/addParent": {
            "post": {
                "tags": ["student"],
                "summary": "Add a parent",
                "description": "Add a parent email to a student's account so that they get notified of changes in the student's performance",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "the addParent object",
                    "schema": {
                        "$ref": "#/definitions/addParentBody"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successfully added the new parent email",
                        "schema": {
                            "$ref": "#/definitions/addParent200Response"
                        }
                    },
                    "404":{
                        "description": "Student not found",
                        "schema": {
                            "$ref": "#/definitions/404IdResponse"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/getMarks": {
            "post": {
                "tags": ["student"],
                "summary": "Get all the marks of a student",
                "description": "Returns the marks of a student",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "description": "Request body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/getMarksBody"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Returns the required results",
                        "schema": {
                            "$ref": "#/definitions/getMarks200Response"
                        }
                    },
                    "404": {
                        "description": "All the sent _ids don't exist in the database",
                        "schema": {
                            "$ref": "#/definitions/404IdResponse"
                        }
                    },
                    "500": {
                        "description": "Something went wrong on the server side",
                        "schema": {
                            "$ref": "#/definitions/500Response"
                        }
                    }
                }
            }
        },
        "/student/getSummary": {
            "post": {
                "tags": ["student"],
                "summary": "Get Summary of Lessons and Performance to make a chart or do statistics",
                "description": "Returns data about all of the subjects that a teacher teaches and the summary of all the students' performance in said subjects",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [{
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "Body containing the required info",
                    "schema": {
                        "$ref": "#/definitions/getSummaryBody"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the summary",
                        "schema": {
                            "$ref": "#/definitions/getSummary200Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request: The subjects array is empty or not present",
                        "schema": {
                            "$ref": "#/definitions/getSummary400Response"
                        }
                    }
                }

            }
        },
...
    },
    "definitions": {
...
        "newStudentBody": {
            "type": "object",
            "required": ["names", "code", "class", "email"],
            "properties": {
                "names": {
                    "type": "string",
                    "example":"ISHIMWE Vainqueur" 
                },
                "code": {
                    "type": "",
                    "example": "RCA0231KIS"
                },
                "class": {
                    "type": "object",
                    "example": {
                        "year": 1,
                        "class": "E"
                    }
                },
                "email": {
                    "type": "string",
                    "example": "ishimvainqueur@gmail.com"
                }
            }
        },
        "newStudent200Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "#Success"
                }
            }
        },
...
        "500Response":{
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "#InternalServerError"
                },
                "message": {
                    "type": "object",
                    "example": {
                        "error": "MongoError",
                        "message": "Connection timeout"
                    }
                }
            }
        },
...
        "getRecordsYearQuery": {
            "type": "number",
            "example": 1
        },
...
    }
}